projects/Shadow-Removal/requirements.txt
cycler==0.11.0
flake8==4.0.1
fonttools==4.29.1
imageio==2.16.0
joblib==1.1.0
kiwisolver==1.3.2
matplotlib==3.5.1
mccabe==0.6.1
networkx==2.6.3
numpy==1.21.5
packaging==21.3
Pillow==9.0.1
pycodestyle==2.8.0
pyflakes==2.4.0
pyparsing==3.0.7
python-dateutil==2.8.2
PyWavelets==1.2.0
rise==5.7.0
scikit-image==0.19.2
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
threadpoolctl==3.1.0
tifffile==2021.11.2
tqdm==4.63.0